[#5 fix] enum Month 외부로
